test/requirements.txt
requests
opencv-contrib-python
numpy
matplotlib
Pillow